+++ angular/playground/projects/ng-router-store/src/public-api.ts
@@ -1,0 +1,4 @@
+export * from './lib/router.module';
+export * from './lib/router.query';
+export * from './lib/router.service';
+export * from './lib/router.store';