+++ angular/playground/src/app/auth/auth.guard.ts
@@ -11,6 +11,7 @@ export class AuthGuard {
   constructor(private router: Router, private authQuery: AuthQuery) {}
 
   canActivate(): Observable<boolean> {
+    // For sync storage
     return this.authQuery.isLoggedIn$.pipe(
       map(isAuth => {
         if (isAuth) {
@@ -21,5 +22,20 @@ export class AuthGuard {
       }),
       take(1)
     );
+
+    // For async storage
+    // return combineLatest([
+    //   this.authQuery.isLoggedIn$,
+    //   selectPersistStateInit(),
+    // ]).pipe(
+    //   map(([isAuth]) => {
+    //     if(isAuth) {
+    //       return true;
+    //     }
+    //     this.router.navigateByUrl('login');
+    //     return false;
+    //   }),
+    //   take(1)
+    // );
   }
 }