+++ CHANGELOG.md
@@ -2,6 +2,43 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [4.14.1](https://github.com/datorama/akita/compare/v4.14.0...v4.14.1) (2019-11-17)
+
+
+### Bug Fixes
+
+* 🐛 persist state dynamic stores ([01efe1f](https://github.com/datorama/akita/commit/01efe1f))
+
+
+
+## [4.14.0](https://github.com/datorama/akita/compare/v4.13.0...v4.14.0) (2019-11-17)
+
+
+### Features
+
+* 🎸 add stores whitelist to devtools ([12e6a76](https://github.com/datorama/akita/commit/12e6a76))
+
+
+
+## [4.13.0](https://github.com/datorama/akita/compare/v4.12.1...v4.13.0) (2019-11-15)
+
+
+### Features
+
+* 🎸 add persist state init to persist state ([b3df8ba](https://github.com/datorama/akita/commit/b3df8ba)), closes [#329](https://github.com/datorama/akita/issues/329)
+
+
+
+### [4.12.1](https://github.com/datorama/akita/compare/v4.12.0...v4.12.1) (2019-11-14)
+
+
+### Bug Fixes
+
+* 🐛 pre add entity type can be any ([85b5ba1](https://github.com/datorama/akita/commit/85b5ba1))
+* **with-transaction:** improve typings ([#327](https://github.com/datorama/akita/issues/327)) ([eb91f89](https://github.com/datorama/akita/commit/eb91f89))
+
+
+
 ## [4.12.0](https://github.com/datorama/akita/compare/v4.11.2...v4.12.0) (2019-11-11)